exec: Extract backend implementation to its own package
This change should allow explicit usage of the `exec` backend by
clients.

It paves the way for the libnftables backend.

The change is backward compatible and places the `exec` as the default
backend (as it is today).
