+++ nft/exec.go
@@ -18,74 +18,3 @@
  */
 
 package nft
-
-import (
-	"bytes"
-	"fmt"
-	"os/exec"
-	"strings"
-)
-
-const (
-	cmdBin     = "nft"
-	cmdFile    = "-f"
-	cmdJSON    = "-j"
-	cmdList    = "list"
-	cmdRuleset = "ruleset"
-	cmdStdin   = "-"
-)
-
-// ReadConfig loads the nftables configuration from the system and
-// returns it as a nftables config structure.
-// The system is expected to have the `nft` executable deployed and nftables enabled in the kernel.
-func ReadConfig() (*Config, error) {
-	stdout, err := execCommand(nil, cmdJSON, cmdList, cmdRuleset)
-	if err != nil {
-		return nil, err
-	}
-
-	config := NewConfig()
-	if err := config.FromJSON(stdout.Bytes()); err != nil {
-		return nil, fmt.Errorf("failed to list ruleset: %v", err)
-	}
-
-	return config, nil
-}
-
-// ApplyConfig applies the given nftables config on the system.
-// The system is expected to have the `nft` executable deployed and nftables enabled in the kernel.
-func ApplyConfig(c *Config) error {
-	data, err := c.ToJSON()
-	if err != nil {
-		return err
-	}
-
-	if _, err := execCommand(data, cmdJSON, cmdFile, cmdStdin); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func execCommand(input []byte, args ...string) (*bytes.Buffer, error) {
-	cmd := exec.Command(cmdBin, args...)
-
-	var stdout, stderr bytes.Buffer
-	cmd.Stderr = &stderr
-	cmd.Stdout = &stdout
-
-	if input != nil {
-		var stdin bytes.Buffer
-		stdin.Write(input)
-		cmd.Stdin = &stdin
-	}
-
-	if err := cmd.Run(); err != nil {
-		return nil, fmt.Errorf(
-			"failed to execute %s %s: %v stdin:'%s' stdout:'%s' stderr:'%s'",
-			cmd.Path, strings.Join(cmd.Args, " "), err, string(input), stdout.String(), stderr.String(),
-		)
-	}
-
-	return &stdout, nil
-}
